requirements.txt
# # Core Libraries
transformers
transformers[sentencepiece]
torch
torchvision
torchaudio
datasets
sacrebleu
streamlit 
accelerate
rouge_score 
py7zr
pandas
nltk
tqdm
PyYAML
matplotlib
notebook
boto3
ollama
mypy-boto3-s3
python-box
ensure
uvicorn
Jinja2
asyncio
typing
evaluate
starlette
langchain
langchain-core
langchain_community
langchain-pinecone
faiss-cpu
bitsandbytes
sentence_transformers
sse_starlette
fastapi
ipykernel
PyMuPDF
unstructured
unstructured[pdf]


# pip install torch==2.6.0+cu118 torchvision==0.17.0+cu118 torchaudio==2.6.0 --index-url https://download.pytorch.org/whl/cu118
